backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase>=2.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic>=2.9.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0
httpx>=0.26.0,<0.28.0
websockets>=11.0,<12.0

# Data processing (compatible versions)
pandas>=1.5.0,<2.2.0
numpy>=1.24.0,<2.0.0

# ML dependencies (compatible versions)
torch>=1.13.0,<2.8.0
torchvision>=0.14.0,<0.23.0
scikit-learn>=1.2.0,<1.4.0
matplotlib>=3.6.0,<3.9.0
joblib>=1.2.0,<1.4.0

# Database migrations
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
